Reorganize sqltraced package
